Rename SnsTopicByName to SnsPublisher
