+++ JustSaying.AwsTools/IPublisher.cs
@@ -1,0 +1,8 @@
+namespace JustSaying.AwsTools
+{
+    public interface IPublisher
+    {
+        void Publish(string subject, string message);
+        void Configure();
+    }
+}
